001-encrypt_password_by_rsa/pyproject.toml
[tool.poetry]
name = "001-rsa_demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
APIFlask = "^1.1.3"
pycryptodome = "^3.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
